extract mocked data into a single object
