+++ nsdb-cluster/src/test/scala/io/radicalbit/nsdb/cluster/coordinator/ReadCoordinatorTemporalAggregatedStatementsSpec.scala
@@ -16,133 +16,15 @@
 
 package io.radicalbit.nsdb.cluster.coordinator
 
-import akka.pattern.ask
-import akka.util.Timeout
-import io.radicalbit.nsdb.cluster.actor.MetricsDataActor.AddRecordToLocation
-import io.radicalbit.nsdb.cluster.coordinator.MetadataCoordinator.commands.AddLocations
+import io.radicalbit.nsdb.cluster.coordinator.mockedData.MockedData._
 import io.radicalbit.nsdb.common.protocol._
 import io.radicalbit.nsdb.common.statement._
-import io.radicalbit.nsdb.model.{Location, TimeContext}
+import io.radicalbit.nsdb.model.TimeContext
 import io.radicalbit.nsdb.protocol.MessageProtocol.Commands._
 import io.radicalbit.nsdb.protocol.MessageProtocol.Events._
 
-import scala.concurrent.Await
 
-object TemporalDoubleMetric {
-
-  val name = "temporalDoubleMetric"
-
-  val recordsShard1: Seq[Bit] = Seq(
-    Bit(150000, 2.5, Map("surname" -> "Doe"), Map("name" -> "John")),
-    Bit(120000, 3.5, Map("surname" -> "Doe"), Map("name" -> "John"))
-  )
-
-  val recordsShard2: Seq[Bit] = Seq(
-    Bit(90000, 5.5, Map("surname" -> "Doe"), Map("name" -> "John")),
-    Bit(60000, 7.5, Map("surname" -> "Doe"), Map("name" -> "Bill")),
-    Bit(30000, 4.5, Map("surname" -> "Doe"), Map("name" -> "Frank")),
-    Bit(0, 1.5, Map("surname"     -> "Doe"), Map("name" -> "Frankie"))
-  )
-
-  val testRecords: Seq[Bit] = recordsShard1 ++ recordsShard2
-
-}
-
-object TemporalLongMetric {
-
-  val name = "temporalLongMetric"
-
-  val recordsShard1: Seq[Bit] = Seq(
-    Bit(150000L, 2L, Map("surname" -> "Doe"), Map("name" -> "John", "age" -> 15L, "height" -> 30.5)),
-    Bit(120000L, 3L, Map("surname" -> "Doe"), Map("name" -> "John", "age" -> 20L, "height" -> 30.5))
-  )
-
-  val recordsShard2: Seq[Bit] = Seq(
-    Bit(90000L, 5L, Map("surname" -> "Doe"), Map("name" -> "John", "age"    -> 15L, "height" -> 30.5)),
-    Bit(60000L, 7L, Map("surname" -> "Doe"), Map("name" -> "Bill", "age"    -> 15L, "height" -> 31.0)),
-    Bit(30000L, 4L, Map("surname" -> "Doe"), Map("name" -> "Frank", "age"   -> 15L, "height" -> 32.0)),
-    Bit(0L, 1L, Map("surname"     -> "Doe"), Map("name" -> "Frankie", "age" -> 15L, "height" -> 32.0))
-  )
-
-  val testRecords: Seq[Bit] = recordsShard1 ++ recordsShard2
-}
-
-class ReadCoordinatorTemporalAggregatedStatementsSpec extends AbstractReadCoordinatorSpec {
-
-  override def beforeAll: Unit = {
-    import scala.concurrent.duration._
-    implicit val timeout: Timeout = Timeout(5.second)
-
-    Await.result(readCoordinatorActor ? SubscribeMetricsDataActor(metricsDataActor, "node1"), 10 seconds)
-
-    val location1 = Location(_: String, "node1", 100000, 190000)
-    val location2 = Location(_: String, "node1", 0, 90000)
-
-    //drop metrics
-    Await.result(
-      metricsDataActor ? DropMetricWithLocations(db,
-                                                 namespace,
-                                                 TemporalLongMetric.name,
-                                                 Seq(location1(TemporalLongMetric.name),
-                                                     location2(TemporalLongMetric.name))),
-      10 seconds
-    )
-    Await.result(
-      metricsDataActor ? DropMetricWithLocations(db,
-                                                 namespace,
-                                                 TemporalDoubleMetric.name,
-                                                 Seq(location1(TemporalDoubleMetric.name),
-                                                     location2(TemporalDoubleMetric.name))),
-      10 seconds
-    )
-
-    Await.result(schemaCoordinator ? UpdateSchemaFromRecord(db,
-                                                            namespace,
-                                                            TemporalLongMetric.name,
-                                                            TemporalLongMetric.testRecords.head),
-                 10 seconds)
-    Await.result(schemaCoordinator ? UpdateSchemaFromRecord(db,
-                                                            namespace,
-                                                            TemporalDoubleMetric.name,
-                                                            TemporalDoubleMetric.testRecords.head),
-                 10 seconds)
-
-    Await.result(
-      metadataCoordinator ? AddLocations(db,
-                                         namespace,
-                                         Seq(location1(TemporalLongMetric.name), location1(TemporalDoubleMetric.name))),
-      10 seconds)
-
-    TemporalLongMetric.recordsShard1
-      .foreach(r => {
-        Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location1(TemporalLongMetric.name)),
-                     10 seconds)
-      })
-    TemporalDoubleMetric.recordsShard1
-      .foreach(r => {
-        Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location1(TemporalDoubleMetric.name)),
-                     10 seconds)
-      })
-
-    Await.result(
-      metadataCoordinator ? AddLocations(db,
-                                         namespace,
-                                         Seq(location2(TemporalLongMetric.name), location2(TemporalDoubleMetric.name))),
-      10 seconds)
-
-    TemporalLongMetric.recordsShard2
-      .foreach(r => {
-        Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location2(TemporalLongMetric.name)),
-                     10 seconds)
-      })
-    TemporalDoubleMetric.recordsShard2
-      .foreach(r => {
-        Await.result(metricsDataActor ? AddRecordToLocation(db, namespace, r, location2(TemporalDoubleMetric.name)),
-                     10 seconds)
-      })
-
-    expectNoMessage(indexingInterval)
-  }
+class ReadCoordinatorTemporalAggregatedStatementsSpec extends AbstractTemporalReadCoordinatorSpec {
 
   "ReadCoordinator" when {
